create AstUtils
